# TINKERPOP-2813 Make gremlin-driver more optimistic about host availability