--- a/docs/src/upgrade/release-3.5.x.asciidoc
+++ b/docs/src/upgrade/release-3.5.x.asciidoc
@@ -62,7 +62,7 @@ The consequences for marking a host dead are fairly severe as the connection poo
 in an orderly fashion so as to allow for existing requests to complete if at all possible, it immediately blocks new
 requests producing an immediate `NoHostAvailableException` (assuming there are actually no other hosts to route
 requests to). There is then some delay to reconnect to the server and re-establish each of the connections in the pool,
-which might take some time especially if there was a large pool. If an application were sending hundreds of request
+which might take some time especially if there was a large pool. If an application were sending hundreds of requests
 a second, it would quickly equate to hundreds of exceptions a second, filling logs quickly and making it hard to
 decipher the original root cause from that initial inability to simply borrow a connection.
 
@@ -76,14 +76,19 @@ state for future requests and if it fails it is finally marked as dead and thus 
 
 If a host is marked unavailable, the driver, like before, will try to route requests to other configured hosts that
 are healthy. If there are no such other hosts and all are unhealthy, the driver will now make an immediate attempt to
-route the request to a random unavailable host on the off chance it will come back online and the connection pool thus
+route the request to a random unavailable host on the chance it will come back online and the connection pool thus
 re-established. So, rather than a `NoHostAvailableException` flood, the user would instead see a `TimeoutException`
 per request made.
 
 Since the driver was already throwing `NoHostAvailableException` and `TimeoutException`, there shouldn't be too
 many code changes required on upgrade to this version. The main difference to consider is that
-`NoHostAvailableException` should appear less often, so any code that mitigated against a flood of these exceptions,
-such defenses might not be necessary anymore.
+`NoHostAvailableException` should appear less often, so any code that mitigated against a flood of these exceptions
+might not be necessary anymore. One should also consider that if code was relying on the asynchronous
+properties that `Client.submitAsync()` implies in its name, that code may now see more blocking behavior than
+before. In doing this refactoring, it was noted that the connection borrowing behavior has never behaved in a truly
+asynchronous manner and only the fast `NoHostAvailableException` aspect of that method behavior really contributed to
+an asynchronous looking call for failed connection borrowing. With the fast `NoHostAvailableException` gone, it may be
+more likely now to note blocking behavior when a connection cannot be obtained.
 
 See: link:https://issues.apache.org/jira/browse/TINKERPOP-2813[TINKERPOP-2813]