… gremlin-driver/src/main/java/org/apache/tinkerpop/gremlin/driver/Connection.java
… gremlin-driver/src/main/java/org/apache/tinkerpop/gremlin/driver/ConnectionFactory.java
… gremlin-driver/src/main/java/org/apache/tinkerpop/gremlin/driver/Host.java
… gremlin-driver/src/main/java/org/apache/tinkerpop/gremlin/driver/Settings.java
… gremlin-driver/src/main/java/org/apache/tinkerpop/gremlin/driver/exception/NoHostAvailableException.java
… gremlin-driver/src/test/java/org/apache/tinkerpop/gremlin/driver/ClientTest.java
… gremlin-driver/src/test/java/org/apache/tinkerpop/gremlin/driver/WebSocketClientBehaviorIntegrateTest.java
… gremlin-server/src/test/java/org/apache/tinkerpop/gremlin/driver/ClientConnectionIntegrateTest.java
… gremlin-server/src/test/java/org/apache/tinkerpop/gremlin/server/AbstractGremlinServerIntegrationTest.java
… gremlin-server/src/test/java/org/apache/tinkerpop/gremlin/server/GremlinDriverIntegrateTest.java