https://issues.apache.org/jira/browse/TINKERPOP-2813

Summarized the major points of the change in the Upgrade Docs. Might be best to start there if reviewing this. Made some other adjustments to log better data for debugging purposes with the `ConnectionPool`.  Ended up making a number of adjustments to the thread pools to which various jobs were being submitted. Untangling that led to the addition of two pools which helped bring back parallel host initialization and got rid of random usage of the fork/join pool. I sense that something could be done to eliminate some of these pools but I believe it will come at the cost of a much larger refactoring effort to untangle how some of these jobs get scheduled up. I dunno.....not super pleased with it but gotta go pencils down at some point.

there've been many days of continuous testing on these changes at this point from various folks and all results have been favorable. calling out @kenhuuu specifically, he did some basic performance testing and profiling and didn't find any problems to note:

```text
Test Case:
    2450 short queries (return in milliseconds)
    1280 mix of short and medium queries

Setup:
    m5.2xlarge
    3.5.4 Gremlin Server with gremlin-server.yaml
    Defaults for connection pooling and in flight per connection settings for Java driver

Results for Three Run Average.
    3.5.4:
        Short: 29912ms
        Mixed: 65689ms

    TINKERPOP-2813 branch:
        Short: 30161ms
        Mixed: 65422ms
```