@kenhuuu 2022-12-14
LGTM. I've tested this and it greatly improves the handling of temporary errors.