@spmallette 2022-12-15
@FlorianHockmann Thanks for the review. I addressed your points. More refinements to the documentation are coming I think, as @kenhuuu will be adding a follow-on PR with more tests. I'll follow up that with a few more documentation adjustments.